refactor(logging): move all logging code into internal package
Consolidate LoggingConfig, InitLog, ReplaceAttr, CleanSourcePath, and
LevelTrace into internal/log.go. This keeps all logging concerns in one
package, shared by both cmd/litestream and internal/testingutil.
