+++ cmd/litestream/replicate.go
@@ -20,6 +20,7 @@ import (
 	"github.com/benbjohnson/litestream/abs"
 	"github.com/benbjohnson/litestream/file"
 	"github.com/benbjohnson/litestream/gs"
+	"github.com/benbjohnson/litestream/internal"
 	"github.com/benbjohnson/litestream/nats"
 	"github.com/benbjohnson/litestream/oss"
 	"github.com/benbjohnson/litestream/s3"
@@ -95,7 +96,7 @@ func (c *ReplicateCommand) ParseFlags(_ context.Context, args []string) (err err
 			if c.Config.Logging.Stderr {
 				logOutput = os.Stderr
 			}
-			initLog(logOutput, c.Config.Logging.Level, c.Config.Logging.Type, c.Config.Logging.Source)
+			internal.InitLog(logOutput, c.Config.Logging.Level, c.Config.Logging.Type, c.Config.Logging.Source)
 		}
 
 	case 1:
@@ -117,7 +118,7 @@ func (c *ReplicateCommand) ParseFlags(_ context.Context, args []string) (err err
 			os.Setenv("LOG_LEVEL", *logLevelFlag)
 		}
 		c.Config.Logging.Level = logLevel
-		initLog(os.Stdout, logLevel, "text", false)
+		internal.InitLog(os.Stdout, logLevel, "text", false)
 
 		dbConfig := &DBConfig{
 			Path:                 fs.Arg(0),
