+++ cmd/litestream/restore.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/benbjohnson/litestream"
+	"github.com/benbjohnson/litestream/internal"
 )
 
 // RestoreCommand represents a command to restore a database from a backup.
@@ -40,7 +41,7 @@ func (c *RestoreCommand) Run(ctx context.Context, args []string) (err error) {
 		return fmt.Errorf("too many arguments")
 	}
 
-	initLog(os.Stdout, "INFO", "text", false)
+	internal.InitLog(os.Stdout, "INFO", "text", false)
 
 	// When follow mode is enabled, set up signal handling so Ctrl+C stops
 	// the follow loop cleanly.
