+++ internal/internal.go
@@ -2,29 +2,13 @@ package internal
 
 import (
 	"io"
-	"log/slog"
 	"os"
-	"path"
-	"runtime/debug"
-	"strings"
 	"syscall"
 
 	"github.com/pierrec/lz4/v4"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 )
-
-var modulePrefix string
-var moduleLastSegment string
-
-func init() {
-	if info, ok := debug.ReadBuildInfo(); ok && info.Main.Path != "" {
-		modulePrefix = info.Main.Path + "/"
-		moduleLastSegment = path.Base(info.Main.Path)
-	}
-}
-
-const LevelTrace = slog.LevelDebug - 4
 
 // ReadCloser wraps a reader to also attach a separate closer.
 type ReadCloser struct {
@@ -161,47 +145,6 @@ func MkdirAll(path string, fi os.FileInfo) error {
 	}
 	_ = os.Chown(path, uid, gid)
 	return nil
-}
-
-func CleanSourcePath(src *slog.Source) {
-	if modulePrefix == "" {
-		return
-	}
-
-	if idx := strings.Index(src.File, modulePrefix); idx >= 0 {
-		src.File = src.File[idx+len(modulePrefix):]
-		return
-	}
-
-	moduleName := strings.TrimSuffix(modulePrefix, "/")
-	if idx := strings.Index(src.File, moduleName+"@"); idx >= 0 {
-		rest := src.File[idx+len(moduleName):]
-		if slashIdx := strings.Index(rest, "/"); slashIdx >= 0 {
-			src.File = rest[slashIdx+1:]
-		}
-		return
-	}
-
-	if moduleLastSegment != "" {
-		sep := "/" + moduleLastSegment + "/"
-		if idx := strings.LastIndex(src.File, sep); idx >= 0 {
-			src.File = src.File[idx+len(sep):]
-		}
-	}
-}
-
-func ReplaceAttr(groups []string, a slog.Attr) slog.Attr {
-	if a.Key == slog.LevelKey && a.Value.Any() == LevelTrace {
-		a.Value = slog.StringValue("TRACE")
-	}
-
-	if a.Key == slog.SourceKey {
-		if src, ok := a.Value.Any().(*slog.Source); ok {
-			CleanSourcePath(src)
-		}
-	}
-
-	return a
 }
 
 // Shared replica metrics.
